feat: Add new Formula for Envctl
This commit introduces a new Homebrew formula for the Envctl project.
The Envctl tool facilitates outputting date formats to the console. The
formula is added to ensure ease of installation and manage dependencies
on 'bun' and 'make'. Bundler and make are utilized during the install
process, with the built binary being placed into the bin directory. The
test block is currently set to fail as a placeholder for future tests.